requirements.txt
flask
flask_admin
flask_login
flask_restful
flask_wtf
python-dotenv
sqlalchemy
sqlalchemy_serializer
elasticsearch
tqdm
requests
BeautifulSoup
sphinx_revealjs